requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jinja2>=3.1.0
click>=8.1.0
colorama>=0.4.6
tabulate>=0.9.0
chardet>=5.0.0
google-cloud-bigquery>=3.0.0 
db-dtypes>=1.0.0